My implementation of the beast attack
The file encrypt is provided compiled, it encrypts some messege where we specify a prefix of length up to 8 bytes. We have to discover how the IV is generated, predict it and use that to eventually decrypt the message. The function callEncrypt was provided and I left it mostly untouched.